Removed client-certificates commands to focus scope of this PR
