+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/SystemCommandLine/Commands.xml
@@ -19,19 +19,6 @@
       <SeeAlso Title="Package source sections in NuGet.config files" Url="/nuget/reference/nuget-config-file#package-source-sections" />
       <SeeAlso Title="sources command (nuget.exe)" Url="/nuget/reference/cli-reference/cli-ref-sources" />
     </Noun>
-    <Noun Name="client-cert" Version="3.1.200"
-       Description="adds a new client certificate to your NuGet configuration files.">
-      <SingleValueOption Name="package-source" Shortcut="s" Help="Option_PackageSource"/>
-      <SingleValueOption Name="path" Help="Option_Path"/>
-      <SingleValueOption Name="password" Help="Option_Password"/>
-      <SwitchOption Name="store-password-in-clear-text" Help="Option_StorePasswordInClearText"/>
-      <SingleValueOption Name="store-location" Help="Option_StoreLocation"/>
-      <SingleValueOption Name="store-name" Help="Option_StoreName"/>
-      <SingleValueOption Name="find-by" Help="Option_FindBy"/>
-      <SingleValueOption Name="find-value" Help="Option_FindValue"/>
-      <SwitchOption Name="force" Shortcut="f" Help="Option_Force"/>
-      <SingleValueOption Name="configfile" Help="Option_ConfigFile" />
-    </Noun>
   </Verb>
   <Verb Name="disable">
     <Noun Name="source" Version="3.1.200"
@@ -62,10 +49,6 @@
       <SeeAlso Title="Package source sections in NuGet.config files" Url="/nuget/reference/nuget-config-file#package-source-sections" />
       <SeeAlso Title="sources command (nuget.exe)" Url="/nuget/reference/cli-reference/cli-ref-sources" />
     </Noun>
-    <Noun Name="client-cert" Version="3.1.200"
-       Description="lists all configured client certificates.">
-      <SingleValueOption Name="configfile" Help="Option_ConfigFile" />
-    </Noun>
   </Verb>
   <Verb Name="remove">
     <Noun Name="source" Version="3.1.200"
@@ -75,11 +58,6 @@
       <Example Title="Remove a source with name of `mySource`:" Command="dotnet nuget remove source mySource" />
       <SeeAlso Title="Package source sections in NuGet.config files" Url="/nuget/reference/nuget-config-file#package-source-sections" />
       <SeeAlso Title="sources command (nuget.exe)" Url="/nuget/reference/cli-reference/cli-ref-sources" />
-    </Noun>
-    <Noun Name="client-cert" Version="3.1.200"
-       Description="removes an existing client certificate configuration from your NuGet configuration files.">
-      <SingleValueOption Name="package-source" Shortcut="s" Help="Option_PackageSource"/>
-      <SingleValueOption Name="configfile" Help="Option_ConfigFile" />
     </Noun>
   </Verb>
   <Verb Name="update">
@@ -95,19 +73,6 @@
       <Example Title="Update a source with name of `mySource`:" Command="dotnet nuget update source mySource --source c:\packages" />
       <SeeAlso Title="Package source sections in NuGet.config files" Url="/nuget/reference/nuget-config-file#package-source-sections" />
       <SeeAlso Title="sources command (nuget.exe)" Url="/nuget/reference/cli-reference/cli-ref-sources" />
-    </Noun>
-    <Noun Name="client-cert" Version="3.1.200"
-       Description="updates an existing client certificate in your NuGet configuration files.">
-      <SingleValueOption Name="package-source" Shortcut="s" Help="Option_PackageSource"/>
-      <SingleValueOption Name="path" Help="Option_Path"/>
-      <SingleValueOption Name="password" Help="Option_Password"/>
-      <SwitchOption Name="store-password-in-clear-text" Help="Option_StorePasswordInClearText"/>
-      <SingleValueOption Name="store-location" Help="Option_StoreLocation"/>
-      <SingleValueOption Name="store-name" Help="Option_StoreName"/>
-      <SingleValueOption Name="find-by" Help="Option_FindBy"/>
-      <SingleValueOption Name="find-value" Help="Option_FindValue"/>
-      <SwitchOption Name="force" Shortcut="f" Help="Option_Force"/>
-      <SingleValueOption Name="configfile" Help="Option_ConfigFile" />
     </Noun>
   </Verb>
 </Commands>
