+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/SystemCommandLine/CustomBinders.cs
@@ -48,52 +48,6 @@ namespace NuGet.CommandLine.XPlat.Commands
         } // end GetBoundValue method
     } // end class
 
-    internal partial class AddClientCertCustomBinder : BinderBase<AddClientCertArgs>
-    {
-        private readonly Option<string> _packageSource;
-        private readonly Option<string> _path;
-        private readonly Option<string> _password;
-        private readonly Option<bool> _storePasswordInClearText;
-        private readonly Option<string> _storeLocation;
-        private readonly Option<string> _storeName;
-        private readonly Option<string> _findBy;
-        private readonly Option<string> _findValue;
-        private readonly Option<bool> _force;
-        private readonly Option<string> _configfile;
-
-        public AddClientCertCustomBinder(Option<string> packageSource, Option<string> path, Option<string> password, Option<bool> storePasswordInClearText, Option<string> storeLocation, Option<string> storeName, Option<string> findBy, Option<string> findValue, Option<bool> force, Option<string> configfile)
-        {
-            _packageSource = packageSource;
-            _path = path;
-            _password = password;
-            _storePasswordInClearText = storePasswordInClearText;
-            _storeLocation = storeLocation;
-            _storeName = storeName;
-            _findBy = findBy;
-            _findValue = findValue;
-            _force = force;
-            _configfile = configfile;
-        }
-
-        protected override AddClientCertArgs GetBoundValue(BindingContext bindingContext)
-        {
-            var returnValue = new AddClientCertArgs()
-            {
-                PackageSource = bindingContext.ParseResult.GetValueForOption(_packageSource),
-                Path = bindingContext.ParseResult.GetValueForOption(_path),
-                Password = bindingContext.ParseResult.GetValueForOption(_password),
-                StorePasswordInClearText = bindingContext.ParseResult.GetValueForOption(_storePasswordInClearText),
-                StoreLocation = bindingContext.ParseResult.GetValueForOption(_storeLocation),
-                StoreName = bindingContext.ParseResult.GetValueForOption(_storeName),
-                FindBy = bindingContext.ParseResult.GetValueForOption(_findBy),
-                FindValue = bindingContext.ParseResult.GetValueForOption(_findValue),
-                Force = bindingContext.ParseResult.GetValueForOption(_force),
-                Configfile = bindingContext.ParseResult.GetValueForOption(_configfile),
-            };
-            return returnValue;
-        } // end GetBoundValue method
-    } // end class
-
     internal partial class DisableSourceCustomBinder : BinderBase<DisableSourceArgs>
     {
         private readonly Argument<string> _name;
@@ -160,25 +114,6 @@ namespace NuGet.CommandLine.XPlat.Commands
         } // end GetBoundValue method
     } // end class
 
-    internal partial class ListClientCertCustomBinder : BinderBase<ListClientCertArgs>
-    {
-        private readonly Option<string> _configfile;
-
-        public ListClientCertCustomBinder(Option<string> configfile)
-        {
-            _configfile = configfile;
-        }
-
-        protected override ListClientCertArgs GetBoundValue(BindingContext bindingContext)
-        {
-            var returnValue = new ListClientCertArgs()
-            {
-                Configfile = bindingContext.ParseResult.GetValueForOption(_configfile),
-            };
-            return returnValue;
-        } // end GetBoundValue method
-    } // end class
-
     internal partial class RemoveSourceCustomBinder : BinderBase<RemoveSourceArgs>
     {
         private readonly Argument<string> _name;
@@ -195,28 +130,6 @@ namespace NuGet.CommandLine.XPlat.Commands
             var returnValue = new RemoveSourceArgs()
             {
                 Name = bindingContext.ParseResult.GetValueForArgument(_name),
-                Configfile = bindingContext.ParseResult.GetValueForOption(_configfile),
-            };
-            return returnValue;
-        } // end GetBoundValue method
-    } // end class
-
-    internal partial class RemoveClientCertCustomBinder : BinderBase<RemoveClientCertArgs>
-    {
-        private readonly Option<string> _packageSource;
-        private readonly Option<string> _configfile;
-
-        public RemoveClientCertCustomBinder(Option<string> packageSource, Option<string> configfile)
-        {
-            _packageSource = packageSource;
-            _configfile = configfile;
-        }
-
-        protected override RemoveClientCertArgs GetBoundValue(BindingContext bindingContext)
-        {
-            var returnValue = new RemoveClientCertArgs()
-            {
-                PackageSource = bindingContext.ParseResult.GetValueForOption(_packageSource),
                 Configfile = bindingContext.ParseResult.GetValueForOption(_configfile),
             };
             return returnValue;
@@ -254,52 +167,6 @@ namespace NuGet.CommandLine.XPlat.Commands
                 Password = bindingContext.ParseResult.GetValueForOption(_password),
                 StorePasswordInClearText = bindingContext.ParseResult.GetValueForOption(_storePasswordInClearText),
                 ValidAuthenticationTypes = bindingContext.ParseResult.GetValueForOption(_validAuthenticationTypes),
-                Configfile = bindingContext.ParseResult.GetValueForOption(_configfile),
-            };
-            return returnValue;
-        } // end GetBoundValue method
-    } // end class
-
-    internal partial class UpdateClientCertCustomBinder : BinderBase<UpdateClientCertArgs>
-    {
-        private readonly Option<string> _packageSource;
-        private readonly Option<string> _path;
-        private readonly Option<string> _password;
-        private readonly Option<bool> _storePasswordInClearText;
-        private readonly Option<string> _storeLocation;
-        private readonly Option<string> _storeName;
-        private readonly Option<string> _findBy;
-        private readonly Option<string> _findValue;
-        private readonly Option<bool> _force;
-        private readonly Option<string> _configfile;
-
-        public UpdateClientCertCustomBinder(Option<string> packageSource, Option<string> path, Option<string> password, Option<bool> storePasswordInClearText, Option<string> storeLocation, Option<string> storeName, Option<string> findBy, Option<string> findValue, Option<bool> force, Option<string> configfile)
-        {
-            _packageSource = packageSource;
-            _path = path;
-            _password = password;
-            _storePasswordInClearText = storePasswordInClearText;
-            _storeLocation = storeLocation;
-            _storeName = storeName;
-            _findBy = findBy;
-            _findValue = findValue;
-            _force = force;
-            _configfile = configfile;
-        }
-
-        protected override UpdateClientCertArgs GetBoundValue(BindingContext bindingContext)
-        {
-            var returnValue = new UpdateClientCertArgs()
-            {
-                PackageSource = bindingContext.ParseResult.GetValueForOption(_packageSource),
-                Path = bindingContext.ParseResult.GetValueForOption(_path),
-                Password = bindingContext.ParseResult.GetValueForOption(_password),
-                StorePasswordInClearText = bindingContext.ParseResult.GetValueForOption(_storePasswordInClearText),
-                StoreLocation = bindingContext.ParseResult.GetValueForOption(_storeLocation),
-                StoreName = bindingContext.ParseResult.GetValueForOption(_storeName),
-                FindBy = bindingContext.ParseResult.GetValueForOption(_findBy),
-                FindValue = bindingContext.ParseResult.GetValueForOption(_findValue),
-                Force = bindingContext.ParseResult.GetValueForOption(_force),
                 Configfile = bindingContext.ParseResult.GetValueForOption(_configfile),
             };
             return returnValue;
