+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/SystemCommandLine/VerbParsers.cs
@@ -32,14 +32,6 @@ namespace NuGet.CommandLine.XPlat.Commands
 
             AddCmd.AddCommand(SourceCmd);
 
-            // noun sub-command: add client-cert
-            var ClientCertCmd = new Command(name: "client-cert", description: Strings.AddClientCertCommandDescription);
-
-            // Options under sub-command: add client-cert
-            RegisterOptionsForCommandAddClientCert(ClientCertCmd, getLogger);
-
-            AddCmd.AddCommand(ClientCertCmd);
-
             GetLoggerFunction = getLogger;
             CommandExceptionHandler = commandExceptionHandler;
             app.AddCommand(AddCmd);
@@ -100,76 +92,6 @@ namespace NuGet.CommandLine.XPlat.Commands
                 }
                 return Task.FromResult(exitCode);
             }, new AddSourceCustomBinder(source_Argument, name_Option, username_Option, password_Option, storePasswordInClearText_Option, validAuthenticationTypes_Option, configfile_Option));
-        }
-
-        private static void RegisterOptionsForCommandAddClientCert(Command cmd, Func<ILogger> getLogger)
-        {
-            var packageSource_Option = new Option<string>(aliases: new[] { "-s", "--package-source" }, description: Strings.Option_PackageSource)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(packageSource_Option);
-            var path_Option = new Option<string>(name: "--path", description: Strings.Option_Path)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(path_Option);
-            var password_Option = new Option<string>(name: "--password", description: Strings.Option_Password)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(password_Option);
-            var storePasswordInClearText_Option = new Option<bool>(name: "--store-password-in-clear-text", description: Strings.Option_StorePasswordInClearText)
-            {
-                Arity = ArgumentArity.Zero,
-            };
-            cmd.Add(storePasswordInClearText_Option);
-            var storeLocation_Option = new Option<string>(name: "--store-location", description: Strings.Option_StoreLocation)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(storeLocation_Option);
-            var storeName_Option = new Option<string>(name: "--store-name", description: Strings.Option_StoreName)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(storeName_Option);
-            var findBy_Option = new Option<string>(name: "--find-by", description: Strings.Option_FindBy)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(findBy_Option);
-            var findValue_Option = new Option<string>(name: "--find-value", description: Strings.Option_FindValue)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(findValue_Option);
-            var force_Option = new Option<bool>(aliases: new[] { "-f", "--force" }, description: Strings.Option_Force)
-            {
-                Arity = ArgumentArity.Zero,
-            };
-            cmd.Add(force_Option);
-            var configfile_Option = new Option<string>(name: "--configfile", description: Strings.Option_ConfigFile)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(configfile_Option);
-            // Create handler delegate handler for cmd
-            cmd.SetHandler((args) =>
-            {
-                int exitCode;
-                try
-                {
-                    AddClientCertRunner.Run(args, getLogger);
-                    exitCode = 0;
-                }
-                catch (Exception e)
-                {
-                    CommandExceptionHandler(e);
-                    exitCode = 1;
-                }
-                return Task.FromResult(exitCode);
-            }, new AddClientCertCustomBinder(packageSource_Option, path_Option, password_Option, storePasswordInClearText_Option, storeLocation_Option, storeName_Option, findBy_Option, findValue_Option, force_Option, configfile_Option));
         }
     } // end class
 
@@ -306,14 +228,6 @@ namespace NuGet.CommandLine.XPlat.Commands
 
             ListCmd.AddCommand(SourceCmd);
 
-            // noun sub-command: list client-cert
-            var ClientCertCmd = new Command(name: "client-cert", description: Strings.ListClientCertCommandDescription);
-
-            // Options under sub-command: list client-cert
-            RegisterOptionsForCommandListClientCert(ClientCertCmd, getLogger);
-
-            ListCmd.AddCommand(ClientCertCmd);
-
             GetLoggerFunction = getLogger;
             CommandExceptionHandler = commandExceptionHandler;
             app.AddCommand(ListCmd);
@@ -350,31 +264,6 @@ namespace NuGet.CommandLine.XPlat.Commands
                 return Task.FromResult(exitCode);
             }, new ListSourceCustomBinder(format_Option, configfile_Option));
         }
-
-        private static void RegisterOptionsForCommandListClientCert(Command cmd, Func<ILogger> getLogger)
-        {
-            var configfile_Option = new Option<string>(name: "--configfile", description: Strings.Option_ConfigFile)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(configfile_Option);
-            // Create handler delegate handler for cmd
-            cmd.SetHandler((args) =>
-            {
-                int exitCode;
-                try
-                {
-                    ListClientCertRunner.Run(args, getLogger);
-                    exitCode = 0;
-                }
-                catch (Exception e)
-                {
-                    CommandExceptionHandler(e);
-                    exitCode = 1;
-                }
-                return Task.FromResult(exitCode);
-            }, new ListClientCertCustomBinder(configfile_Option));
-        }
     } // end class
 
     internal partial class RemoveVerbParser
@@ -395,14 +284,6 @@ namespace NuGet.CommandLine.XPlat.Commands
             RegisterOptionsForCommandRemoveSource(SourceCmd, getLogger);
 
             RemoveCmd.AddCommand(SourceCmd);
-
-            // noun sub-command: remove client-cert
-            var ClientCertCmd = new Command(name: "client-cert", description: Strings.RemoveClientCertCommandDescription);
-
-            // Options under sub-command: remove client-cert
-            RegisterOptionsForCommandRemoveClientCert(ClientCertCmd, getLogger);
-
-            RemoveCmd.AddCommand(ClientCertCmd);
 
             GetLoggerFunction = getLogger;
             CommandExceptionHandler = commandExceptionHandler;
@@ -440,36 +321,6 @@ namespace NuGet.CommandLine.XPlat.Commands
                 return Task.FromResult(exitCode);
             }, new RemoveSourceCustomBinder(name_Argument, configfile_Option));
         }
-
-        private static void RegisterOptionsForCommandRemoveClientCert(Command cmd, Func<ILogger> getLogger)
-        {
-            var packageSource_Option = new Option<string>(aliases: new[] { "-s", "--package-source" }, description: Strings.Option_PackageSource)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(packageSource_Option);
-            var configfile_Option = new Option<string>(name: "--configfile", description: Strings.Option_ConfigFile)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(configfile_Option);
-            // Create handler delegate handler for cmd
-            cmd.SetHandler((args) =>
-            {
-                int exitCode;
-                try
-                {
-                    RemoveClientCertRunner.Run(args, getLogger);
-                    exitCode = 0;
-                }
-                catch (Exception e)
-                {
-                    CommandExceptionHandler(e);
-                    exitCode = 1;
-                }
-                return Task.FromResult(exitCode);
-            }, new RemoveClientCertCustomBinder(packageSource_Option, configfile_Option));
-        }
     } // end class
 
     internal partial class UpdateVerbParser
@@ -490,14 +341,6 @@ namespace NuGet.CommandLine.XPlat.Commands
             RegisterOptionsForCommandUpdateSource(SourceCmd, getLogger);
 
             UpdateCmd.AddCommand(SourceCmd);
-
-            // noun sub-command: update client-cert
-            var ClientCertCmd = new Command(name: "client-cert", description: Strings.UpdateClientCertCommandDescription);
-
-            // Options under sub-command: update client-cert
-            RegisterOptionsForCommandUpdateClientCert(ClientCertCmd, getLogger);
-
-            UpdateCmd.AddCommand(ClientCertCmd);
 
             GetLoggerFunction = getLogger;
             CommandExceptionHandler = commandExceptionHandler;
@@ -559,76 +402,6 @@ namespace NuGet.CommandLine.XPlat.Commands
                 }
                 return Task.FromResult(exitCode);
             }, new UpdateSourceCustomBinder(name_Argument, source_Option, username_Option, password_Option, storePasswordInClearText_Option, validAuthenticationTypes_Option, configfile_Option));
-        }
-
-        private static void RegisterOptionsForCommandUpdateClientCert(Command cmd, Func<ILogger> getLogger)
-        {
-            var packageSource_Option = new Option<string>(aliases: new[] { "-s", "--package-source" }, description: Strings.Option_PackageSource)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(packageSource_Option);
-            var path_Option = new Option<string>(name: "--path", description: Strings.Option_Path)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(path_Option);
-            var password_Option = new Option<string>(name: "--password", description: Strings.Option_Password)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(password_Option);
-            var storePasswordInClearText_Option = new Option<bool>(name: "--store-password-in-clear-text", description: Strings.Option_StorePasswordInClearText)
-            {
-                Arity = ArgumentArity.Zero,
-            };
-            cmd.Add(storePasswordInClearText_Option);
-            var storeLocation_Option = new Option<string>(name: "--store-location", description: Strings.Option_StoreLocation)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(storeLocation_Option);
-            var storeName_Option = new Option<string>(name: "--store-name", description: Strings.Option_StoreName)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(storeName_Option);
-            var findBy_Option = new Option<string>(name: "--find-by", description: Strings.Option_FindBy)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(findBy_Option);
-            var findValue_Option = new Option<string>(name: "--find-value", description: Strings.Option_FindValue)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(findValue_Option);
-            var force_Option = new Option<bool>(aliases: new[] { "-f", "--force" }, description: Strings.Option_Force)
-            {
-                Arity = ArgumentArity.Zero,
-            };
-            cmd.Add(force_Option);
-            var configfile_Option = new Option<string>(name: "--configfile", description: Strings.Option_ConfigFile)
-            {
-                Arity = ArgumentArity.ZeroOrOne,
-            };
-            cmd.Add(configfile_Option);
-            // Create handler delegate handler for cmd
-            cmd.SetHandler((args) =>
-            {
-                int exitCode;
-                try
-                {
-                    UpdateClientCertRunner.Run(args, getLogger);
-                    exitCode = 0;
-                }
-                catch (Exception e)
-                {
-                    CommandExceptionHandler(e);
-                    exitCode = 1;
-                }
-                return Task.FromResult(exitCode);
-            }, new UpdateClientCertCustomBinder(packageSource_Option, path_Option, password_Option, storePasswordInClearText_Option, storeLocation_Option, storeName_Option, findBy_Option, findValue_Option, force_Option, configfile_Option));
         }
     } // end class
