+++ test/TestUtilities/Test.Utility/SimpleTestSetup/SimpleTestSettingsContext.cs
@@ -35,12 +35,6 @@ namespace NuGet.Test.Utility
             ConfigPath = path ?? throw new ArgumentNullException(nameof(path));
         }
 
-        public SimpleTestSettingsContext(string path)
-            : this(path, GetEmptyConfig())
-        {
-
-        }
-
         /// <summary>
         /// Save to disk
         /// </summary>
@@ -90,7 +84,8 @@ namespace NuGet.Test.Utility
             var doc = new XDocument();
             var configuration = new XElement(XName.Get("configuration"));
             doc.Add(configuration);
-            _ = GetOrAddSection(doc, "config");
+
+            var config = GetOrAddSection(doc, "config");
             var packageSources = GetOrAddSection(doc, "packageSources");
             var disabledSources = GetOrAddSection(doc, "disabledPackageSources");
             var fallbackFolders = GetOrAddSection(doc, "fallbackPackageFolders");
